Week5: Django-rest-framework/requirements.txt
Django==1.11
djangorestframework==3.7.7
drfdocs==0.0.11
pytz==2018.3
